bot/requirements.txt
python-telegram-bot>=21.0,<22
httpx>=0.27,<1
python-dotenv>=1.0,<2
